requirements.txt
sys
gui
PyQt6.QtWidgets 
math
matplotlib.pyplot
numpy
setuptools
unittest